Testes da PedidoService / outras alterações

+++ src/main/java/com/remeikis/provainfuse/model/Pedido.java
@@ -31,7 +31,7 @@ public class Pedido
     private String nome;
 
     @Column(nullable = false)
-    private float valor;
+    private double valor;
 
     @Column(nullable = false)
     private Integer quantidade;
